requirements.txt
google-generativeai
Pillow
keyboard
deep-translator
pytesseract